[DT-032892] Add Soft Credit Check to Equifax Module
