+++ Decisions.Equifax/PrequalificationOfOne/Request/PrequalificationOfOneRequest.cs
@@ -1,0 +1,13 @@
+﻿using System.Runtime.Serialization;
+using DecisionsFramework.Design.ConfigurationStorage.Attributes;
+using Newtonsoft.Json;
+using Decisions.Equifax.ConsumerCreditReport.Request;
+
+namespace Decisions.Equifax.PrequalificationOfOne.Request
+{
+    public class PrequalificationOfOneRequest : ConsumerCreditReportRequest
+
+    {
+    /*Since the request are the same, extending the Response class of Consumer Credit Report Request*/
+    }
+}
